Create extractor.py
The key terms, in English, to determine state in the extracted text varied in case depending on what generated the receipt. There was better consistency in French so I used the French terms. The digital signature is a hex-encoding of ASN1 format. There's little value in parsing that, for the pain.